requirements.txt
plotly==4.4.1
pandas==0.25.3
beautifulsoup4==4.8.1
python_igraph==0.7.1.post6
